Inline seat billing helper
